refactor(lsp): split lifecycle tests
